requirements.txt
pillow>=8.0.0
click>=8.0.0
pytest>=6.0.0
opencv-python>=4.5.0
dlib>=19.20.0 
